Refactor popper modifier and apply to all drops
