+++ changelog/unreleased/bugfix-drop-menus-limited-vertical-space
@@ -1,6 +1,6 @@
 Bugfix: Drop menus with limited vertical screen space
 
-We've changed the way how drop menus in resource tables and the right sidebar are positioned when only limited vertical screen space is available. They now fully utilize the available vertical screen space.
+We've changed the way how drop menus are positioned when only limited vertical screen space is available. They now fully utilize the available vertical screen space.
 
 https://github.com/owncloud/web/issues/8411
 https://github.com/owncloud/web/issues/8412
