+++ packages/design-system/package.json
@@ -80,7 +80,6 @@
     "npm-run-all": "^4.1.5",
     "onchange": "^7.1.0",
     "optimize-css-assets-webpack-plugin": "^6.0.1",
-    "popper-max-size-modifier": "^0.2.0",
     "postcss-import": "^12.0.1",
     "postcss-loader": "^3.0.0",
     "postcss-safe-parser": "^5.0.2",
@@ -106,13 +105,13 @@
     "typescript": "^4.3.2",
     "url-loader": "^4.1.1",
     "v-calendar": "github:dschmidt/v-calendar#3ce6e3b8afd5491cb53ee811281d5fa8a45b044d",
-    "vue3-gettext": "^2.3.3",
+    "vue": "3.2.45",
     "vue-inline-svg": "^3.1.0",
     "vue-loader": "^15.9.0",
     "vue-select": "4.0.0-beta.6",
     "vue-style-loader": "^4.1.2",
     "vue-styleguidist": "^4.44.2",
-    "vue": "3.2.45",
+    "vue3-gettext": "^2.3.3",
     "web-test-helpers": "workspace:*",
     "webfontloader": "^1.6.28",
     "webpack": "^4.41.6",
