+++ packages/web-app-files/src/components/FilesList/ResourceTiles.vue
@@ -91,7 +91,7 @@ import { Resource, SpaceResource } from 'web-client'
 import { useStore } from 'web-pkg/src/composables'
 import { ImageDimension } from 'web-pkg/src/constants'
 import { createFileRouteOptions } from 'web-pkg/src/helpers/router'
-import { displayPositionedDropdown } from 'web-pkg/src/helpers'
+import { displayPositionedDropdown } from 'web-pkg/src/helpers/contextMenuDropdown'
 import { createLocationSpaces } from 'web-app-files/src/router'
 import ContextMenuQuickAction from 'web-pkg/src/components/ContextActions/ContextMenuQuickAction.vue'
