+++ packages/web-app-files/src/components/SideBar/Shares/Links/DetailsAndEdit.vue
@@ -15,7 +15,6 @@
         class="edit-public-link-role-dropdown"
         :drop-id="`edit-public-link-role-dropdown`"
         :toggle="`#edit-public-link-role-dropdown-toggle-${link.id}`"
-        :popper-options="popperOptions"
         padding-size="small"
         offset="0"
         mode="click"
@@ -100,7 +99,7 @@
       />
       <div v-if="isModifiable">
         <oc-button
-          :id="`edit-public-link-dropdown-toggle-${link.id}`"
+          :id="`edit-public-link-dropdown-toggl-${link.id}`"
           appearance="raw"
           class="edit-drop-trigger"
           :data-testid="`files-link-id-${link.id}-btn-edit`"
@@ -110,7 +109,7 @@
         <oc-drop
           ref="editPublicLinkDropdown"
           :drop-id="`edit-public-link-dropdown`"
-          :toggle="`#edit-public-link-dropdown-toggle-${link.id}`"
+          :toggle="`#edit-public-link-dropdown-toggl-${link.id}`"
           padding-size="small"
           mode="click"
         >
@@ -185,12 +184,7 @@ import {
   ShareRole
 } from 'web-client/src/helpers/share'
 import { defineComponent, inject, PropType } from 'vue'
-import {
-  formatDateFromDateTime,
-  formatRelativeDateFromDateTime,
-  getPopperOptions,
-  getSidebarOffset
-} from 'web-pkg/src/helpers'
+import { formatDateFromDateTime, formatRelativeDateFromDateTime } from 'web-pkg/src/helpers'
 import { Resource, SpaceResource } from 'web-client/src/helpers'
 import { createFileRouteOptions } from 'web-pkg/src/helpers/router'
 
@@ -229,13 +223,7 @@ export default defineComponent({
   },
   emits: ['removePublicLink', 'updateLink'],
   setup() {
-    const popperOptions = getPopperOptions({ topOffset: getSidebarOffset() })
-
-    return {
-      space: inject<Resource>('space'),
-      resource: inject<Resource>('resource'),
-      popperOptions
-    }
+    return { space: inject<Resource>('space'), resource: inject<Resource>('resource') }
   },
   data() {
     return {
