+++ packages/web-pkg/package.json
@@ -19,14 +19,13 @@
     "lodash-es": "^4.17.21",
     "luxon": "^2.4.0",
     "mark.js": "^8.11.1",
-    "popper-max-size-modifier": "^0.2.0",
     "qs": "^6.10.3",
     "uuid": "^9.0.0",
+    "vue-concurrency": "4.0.0",
     "vue-router": "4.1.6",
+    "vue3-gettext": "2.5.0-alpha.1",
     "vuex": "4.1.0",
     "web-client": "workspace:@ownclouders/web-client@*",
-    "web-pkg": "workspace:@ownclouders/web-pkg@*",
-    "vue-concurrency": "4.0.0",
-    "vue3-gettext": "2.5.0-alpha.1"
+    "web-pkg": "workspace:@ownclouders/web-pkg@*"
   }
 }
