+++ packages/web-runtime/package.json
@@ -27,7 +27,6 @@
     "oidc-client-ts": "^2.1.0",
     "owncloud-sdk": "~3.1.0-alpha.3",
     "p-queue": "^6.6.2",
-    "popper-max-size-modifier": "^0.2.0",
     "portal-vue": "3.0.0",
     "postcss-import": "^12.0.1",
     "postcss-url": "^9.0.0",
@@ -40,10 +39,10 @@
     "uuid": "^9.0.0",
     "vue": "3.2.45",
     "vue-concurrency": "4.0.0",
-    "vue3-gettext": "^2.3.3",
     "vue-inline-svg": "3.1.0",
     "vue-router": "4.1.6",
     "vue-select": "4.0.0-beta.6",
+    "vue3-gettext": "^2.3.3",
     "vuex": "4.1.0",
     "web-client": "workspace:@ownclouders/web-client@*",
     "web-pkg": "workspace:@ownclouders/web-pkg@*",
